Better parsing and handling of instrumental tracks. cleanup

+++ internal/ui/visualizers/oscillogram.go
@@ -10,9 +10,9 @@ import (
 var lowerMasks = [7]string{"▇", "▆", "▅", "▄", "▃", "▂", "▁"}
 
 const (
-	oscMinAmp      = 0.005  // resting bar height for the idle gradient line
-	oscDecayActive = 0.82   // band release decay per tick when audio is present
-	oscDecayIdle   = 0.88   // band decay per tick when no audio
+	oscMinAmp      = 0.005          // resting bar height for the idle gradient line
+	oscDecayActive = float32(0.82)  // band release decay per tick when audio is present
+	oscDecayIdle   = float32(0.88)  // band decay per tick when no audio
 )
 
 type Oscillogram struct {
